requirements.txt
alembic==1.4.3
bcrypt==3.2.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
PyJWT==2.0.1
pyngrok==5.0.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-magic-bin==0.4.14
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
